requirements.txt
streamlit
scikit-learn
pandas
numpy
joblib
pyyaml
streamlit-authenticator
requests
beautifulsoup4